+++ extensions/airtable/src/metadata-api.tsx
@@ -5,6 +5,7 @@ import {
   AirtableMetadataApiBaseDetails,
   AirtableMetadataApiBaseListResponse,
   AirtableMetadataApiBaseSchemaResponse,
+  AirtableRecord,
   ErrorResponse,
 } from "./types";
 import { client } from "./oauth-client";
@@ -117,4 +118,27 @@ export async function fetchBaseRecords(baseId: string, tableId: string) {
 
   const result = (await response.json()) as AirtableBaseRecordsResponse;
   return result.records;
+}
+
+export async function updateBaseRecord(baseId: string, tableId: string, recordId: string, fields: Record<string, string>) {
+  const response = await fetch(`${airtableApiBaseUrl}/v0/${baseId}/${tableId}/${recordId}`, {
+    method: "PATCH",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${(await client.getTokens())?.accessToken}`,
+    },
+    body: JSON.stringify({fields})
+  });
+
+  if (!response.ok) {
+    if (response.headers.get("Content-Type")?.includes("application/json")) {
+      const err = (await response.json()) as ErrorResponse;
+      throw new Error(err.error.message);
+    } else {
+      throw new Error(response.statusText);
+    }
+  }
+
+  const result = (await response.json()) as AirtableRecord;
+  return result;
 }